backend/requirements.txt
# FastAPI and web server
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.21
alembic==1.12.0
psycopg2-binary==2.9.7
aiosqlite==0.19.0

# Authentication and security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# API clients and utilities
requests==2.31.0
aiohttp==3.8.5
python-dateutil==2.8.2
pytz==2023.3

# Background tasks and scheduling
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Encryption and secure storage
cryptography==41.0.3

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.2
httpx==0.25.0

# Logging
structlog==23.1.0 